Split up projects controller into separate files
